requirements.txt
pyodbc==4.0.39
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.2
PyQt6==6.5.2
pytest==7.4.0